tests/wpt/tests/signed-exchange/reporting/sxg-reporting-navigation-cert_verification_error-downgraded.tentative.html

<!DOCTYPE html>
<title>SXG reporting test of sxg.failed downgraded from sxg.cert_verification_error for navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/network-error-logging/support/nel.sub.js"></script>
<script src="../resources/sxg-util.js"></script>
<body>
<script>
const alt_origin = get_host_info().HTTPS_NOTSAMESITE_ORIGIN;
const test_origin = get_host_info().HTTPS_ORIGIN;

nel_iframe_test(async t => {
  await loadResourceWithBasicPolicyInIframe();
  const sxg_url =
      alt_origin +
      '/signed-exchange/resources/sxg/sxg-validity-period-too-long.sxg?navigation';
  const cert_url = test_origin + '/signed-exchange/resources/127.0.0.1.sxg.pem.cbor';
  const message = await openSXGInIframeAndWaitForMessage(t, sxg_url);
  assert_equals(message.location,
                innerURLOrigin() + '/signed-exchange/resources/inner-url.html', 'location');
  assert_true(message.is_fallback, 'is_fallback');
  assert_true(await reportsExist([
    {
      url: sxg_url,
      user_agent: navigator.userAgent,
      type: "network-error",
      body: {
        phase: "sxg",
        // The origin of this SXG's certURL is different from the reporting
        // origin. So the report must be downgraded.
        type: "sxg.failed",
        elapsed_time: 0,
        status_code: 200,
        referrer: location.origin + '/',
        sxg: {
          outer_url: sxg_url,
          inner_url:
              innerURLOrigin() + '/signed-exchange/resources/inner-url.html',
          cert_url: [cert_url]
        }
      },
      metadata: {
        content_type: "application/reports+json",
      },
    },
    {
      url: sxg_url,
      user_agent: navigator.userAgent,
      type: "network-error",
      body: {
        phase: "application",
        type: "ok",
        status_code: 200,
        referrer: location.origin + '/',
      },
      metadata: {
        content_type: "application/reports+json",
      },
    },
  ]));
}, 'SXG reporting test of sxg.failed downgraded from ' +
    'sxg.cert_verification_error for navigation.');
</script>
</body>
